GluonInstance now a top-level class.
