+++ modules/gluon/src/org/lucidj/gluon/GluonReader.java
@@ -18,7 +18,6 @@ package org.lucidj.gluon;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.lucidj.gluon.GluonSerializer.GluonInstance;
 
 import java.io.BufferedReader;
 import java.io.IOException;
